requirements.txt
langchain
openai
streamlit
python-environ
tabulate
pandas 
numpy
faker
flask
openai
python-dotenv
pandas